pyproject.toml
[project]
name = "fini"
version = "0.1.0"
description = "Tiny CLI for hard working ants"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "alexjuda", email = "alexanderjuda@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.44",
    "markdown-it-py>=3.0.0",
    "mdformat>=0.7.22",
]

[project.scripts]
fini = "fini:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "freezegun>=1.5.1",
    "ipdb>=0.13.13",
    "ipython>=9.2.0",
    "pyright>=1.1.400",
    "pytest>=8.3.5",
    "ruff>=0.11.8",
]
